apps/backend/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Validation & Config
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Task Queue
celery>=5.3.0
redis[hiredis]>=5.0.0

# Auth
python-jose[cryptography]>=3.3.0
pyjwt[crypto]>=2.8.0  # For OIDC ID token verification
bcrypt>=4.0.0
authlib>=1.2.0
httpx>=0.25.0  # For OIDC provider communication
itsdangerous>=2.1.0  # For secure session handling

# Storage
aiofiles>=23.0.0
boto3>=1.28.0
aioboto3>=12.0.0

# OCR & PDF
PyMuPDF>=1.23.0  # For PDF text extraction
Pillow>=10.0.0  # For image processing
img2pdf>=0.5.0  # For converting images to PDF

# OCR - Strategy pattern supports both PaddleOCR and EasyOCR
# PaddleOCR: Better accuracy, recommended for x86/AMD64 (set OCR_PROVIDER=paddleocr)
# Pin to 2.x versions due to breaking changes and OneDNN bugs in 3.x
paddlepaddle==2.6.2  # CPU version - last 2.x release
paddleocr<3.0.0  # Stay on 2.x to avoid breaking API changes
# For GPU support, use: paddlepaddle-gpu==2.6.2 instead of paddlepaddle

# EasyOCR: Better compatibility, recommended for ARM64 (set OCR_PROVIDER=easyocr)
easyocr>=1.7.0

# OCR_PROVIDER=auto (default) will auto-detect architecture and choose the best option

# Embeddings & LLM
sentence-transformers>=2.2.0
openai>=1.0.0  # For OpenAI embeddings and LLM (optional, set EMBEDDING_PROVIDER=openai or LLM_PROVIDER=openai)
google-generativeai>=0.3.0  # For Gemini LLM (optional, set LLM_PROVIDER=gemini)
ollama>=0.1.0  # For Ollama LLM (optional, set LLM_PROVIDER=ollama)

# Vector DB
pgvector>=0.2.0

# Email (will add in Phase 6)
# imapclient>=2.3.0

# File Watching
watchdog>=3.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
